update to use same functions for fs in src
