+++ packages/cli/src/scripts/update-template.ts
@@ -1,4 +1,3 @@
-import fs from 'node:fs/promises';
 import path from 'node:path';
 import chalk from 'chalk';
 import * as R from 'ramda';
@@ -10,7 +9,7 @@ import accentColorCategoryJson from '../../../../design-tokens/semantic/modes/ma
   type: 'json',
 };
 import digdirThemeJson from '../../../../design-tokens/themes/digdir.json' with { type: 'json' };
-import { cleanDir } from '../tokens/utils.js';
+import { cleanDir, cp, mkdir, writeFile } from '../utils.js';
 
 const DIRNAME: string = import.meta.dirname || __dirname;
 
@@ -22,17 +21,15 @@ const argsFromToPaths = (path_: string): [string, string] => [
   path.join(TEMPLATE_FILES_PATH, path_),
 ];
 
-const options = { recursive: true };
-
 export const updateTemplates = async () => {
   // Clean template files
   await cleanDir(TEMPLATE_FILES_PATH);
 
   // Copy default files
-  await fs.cp(...argsFromToPaths('primitives/globals.json'), options);
-  await fs.cp(...argsFromToPaths('primitives/modes/size'), options);
-  await fs.cp(...argsFromToPaths('primitives/modes/typography/size'), options);
-  await fs.cp(...argsFromToPaths('semantic/style.json'), options);
+  await cp(...argsFromToPaths('primitives/globals.json'));
+  await cp(...argsFromToPaths('primitives/modes/size'));
+  await cp(...argsFromToPaths('primitives/modes/typography/size'));
+  await cp(...argsFromToPaths('semantic/style.json'));
 
   /*
    * Create template files
@@ -45,8 +42,8 @@ export const updateTemplates = async () => {
     digdirTypographyJson,
   );
   const typographyDir = path.join(TEMPLATE_FILES_PATH, 'primitives/modes/typography/');
-  await fs.mkdir(typographyDir, options);
-  await fs.writeFile(
+  await mkdir(typographyDir);
+  await writeFile(
     path.join(typographyDir, 'typography-template.json'),
     JSON.stringify(typographyTemplate, null, 2).replaceAll('digdir', '<theme>'),
   );
@@ -54,8 +51,8 @@ export const updateTemplates = async () => {
   // semantic/modes/<category>-color/<color>.json
   const categoryColorTemplate = accentColorCategoryJson.color.main;
   const categoryDir = path.join(TEMPLATE_FILES_PATH, 'semantic/modes');
-  await fs.mkdir(categoryDir, options);
-  await fs.writeFile(
+  await mkdir(categoryDir);
+  await writeFile(
     path.join(categoryDir, 'color-template.json'),
     JSON.stringify(categoryColorTemplate, null, 2).replaceAll('color.accent', 'color.<color>'),
   );
@@ -66,13 +63,13 @@ export const updateTemplates = async () => {
     // Remove custom colors as they are defined from by theme
     color: R.omit(['accent', 'neutral', 'brand1', 'brand2', 'brand3'], semanticColorJson.color),
   };
-  await fs.writeFile(
+  await writeFile(
     path.join(TEMPLATE_FILES_PATH, `semantic/color-base-template.json`),
     JSON.stringify(colorBaseFile, null, 2),
   );
 
   const semanticColorTemplate = semanticColorJson.color.accent;
-  await fs.writeFile(
+  await writeFile(
     path.join(TEMPLATE_FILES_PATH, `semantic/color-template.json`),
     JSON.stringify(semanticColorTemplate, null, 2).replaceAll('color.accent', 'color.<color>'),
   );
@@ -83,21 +80,21 @@ export const updateTemplates = async () => {
     ...R.omit(['color'], digdirThemeJson),
   };
 
-  await fs.mkdir(path.join(TEMPLATE_FILES_PATH, 'themes'), options);
-  await fs.writeFile(
+  await mkdir(path.join(TEMPLATE_FILES_PATH, 'themes'));
+  await writeFile(
     path.join(TEMPLATE_FILES_PATH, `themes/theme-base-template.json`),
     JSON.stringify(themeBaseFile, null, 2).replaceAll('digdir', '<theme>'),
   );
 
   const themeColorTemplate = digdirThemeJson.color.accent;
-  await fs.writeFile(
+  await writeFile(
     path.join(TEMPLATE_FILES_PATH, `themes/theme-template.json`),
     JSON.stringify(themeColorTemplate, null, 2).replaceAll('digdir.accent', '<theme>.<color>'),
   );
 
   // WIP
   // // $themes.json
-  // const themesFile = await fs.readFile(path.join(SOURCE_FILES_PATH, '$themes.json'), 'utf-8');
+  // const themesFile = await readFile(path.join(SOURCE_FILES_PATH, '$themes.json'), 'utf-8');
   // const themesTemplate = (JSON.parse(themesFile) as ThemeObject[])
   //   .filter((themeobj) => {
   //     // Remove compact size
@@ -125,10 +122,10 @@ export const updateTemplates = async () => {
   //     return themeobj;
   //   });
 
-  // await fs.writeFile(path.join(TEMPLATE_FILES_PATH, `$themes.json`), stringify(themesTemplate));
+  // await writeFile(path.join(TEMPLATE_FILES_PATH, `$themes.json`), stringify(themesTemplate));
 
   // // $metadata.json
-  // const metadataFile = await fs.readFile(path.join(SOURCE_FILES_PATH, '$metadata.json'), 'utf-8');
+  // const metadataFile = await readFile(path.join(SOURCE_FILES_PATH, '$metadata.json'), 'utf-8');
   // const tokenSetOrderTemplate = (JSON.parse(metadataFile) as { tokenSetOrder: string[] }).tokenSetOrder
   //   .filter((tokenSet) => {
   //     if (endsWithOneOf(['altinn', 'portal', 'uutilsynet'], tokenSet)) {
@@ -143,7 +140,7 @@ export const updateTemplates = async () => {
   //     return tokenSet;
   //   });
 
-  // await fs.writeFile(
+  // await writeFile(
   //   path.join(TEMPLATE_FILES_PATH, `$metadata.json`),
   //   stringify({ tokenSetOrder: tokenSetOrderTemplate }),
   // );
@@ -151,4 +148,4 @@ export const updateTemplates = async () => {
   console.log(chalk.green('Templates updated'));
 };
 
-updateTemplates();
+await updateTemplates();
